change directory structure and divide models
