+++ app/src/main/java/com/github/poscat/rss/viewer/model/Author.kt
@@ -1,0 +1,7 @@
+package com.github.poscat.rss.viewer.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Author (
+    @SerializedName("name") var name: String
+)
