+++ app/src/main/java/com/github/poscat/rss/viewer/model/Channel.kt
@@ -1,4 +1,4 @@
-package com.github.poscat.rss.viewer.DataType
+package com.github.poscat.rss.viewer.model
 
 import com.google.gson.annotations.SerializedName
