+++ app/src/main/java/com/github/poscat/rss/viewer/model/Contents.kt
@@ -1,0 +1,8 @@
+package com.github.poscat.rss.viewer.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Contents (
+    @SerializedName("url") var url: String,
+    @SerializedName("type") var type: String
+)
